Remote Control & SignalR
Reworking the handling of SignalR. Building in a small SignalR client to communicate with the user process instead of a normal TCP server/client.

Already added some test logics that worked before with the small tcp client, but the tcp client caused issues which is why we will use SignalR instead.